+++ R/plotting.R
@@ -10,7 +10,6 @@
 #' @importFrom ggthemes scale_colour_ptol
 #' @importFrom xcms chromatogram
 #' @importFrom magrittr set_names
-#' @importFrom patchwork wrap_plots plot_layout
 #' @importFrom ggplot2 ggplot geom_line theme_bw aes labs theme element_text
 #' @importFrom dplyr bind_rows group_by summarise
 #' @importFrom erah plotChr